requirements.txt
pandas==2.1.1
Pillow==10.0.1
requests==2.31.0
openpyxl==3.1.2
xlrd==2.0.1